requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
docutils==0.19
ffpyplayer==4.4.0
idna==3.4
Kivy==2.1.0
Kivy-examples==2.1.0
Kivy-Garden==0.1.5
Pillow==9.4.0
Pygments==2.14.0
requests==2.28.2
stopwatch.py==2.0.1
urllib3==1.26.14
